refactor: share marketplace data contracts
